requirements.txt
ta
pandas
yfinance
fpdf2
prettytable